Artificer Updoot and Nomadic Chef gaming (#3084)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a variety of new quality of life items to Artificer and merchant
shop. large focus on mobile cookware for nomadic chefs to roam with and
missing bronze prosthetics, along with a new item that allows artificers
to easily attach them to otherwise healthy individuals looking to
surpass the limits of their weak flesh.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Large rp potential for refugee cook roles and upcoming nomad roles for
Desert town. Will tie in with a future PR that allows tents to be broken
down and carried as well so you can have a whole campsite on the move
with you. It also adds bronze prosthetic legs which where missing and
unifies the sprites between the new legs with new arm sprites
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Spawned in as artificer and merchant to test the items and recipes and
so far havent had any issues with this current version. Feel free to
review.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->