Allows ranged weapons to earn xp, Artificer recipe fix, few more cosmetic clothing options added to merchant (#3091)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Just fixes something i missed in my last pr but largely is for allowing
ranged weapons to gain xp now
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
may be too op as its basically a direct copy of sling code for xp so can
be toned down but only gives xp if used on a living mob
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
just trust me bro
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->