fixes a few item sizes (#3092)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes recipe book sizes along with lowering the size of items like the
love potion and the bullet and sling pouches considering they where
never updated for the new inventory system
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
the size of the survival guide prevents items like the mages spell book
for some classes from spawning as it takes up the majority of their bag
space.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
booted it up and testing on a private server
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->