Adds Grain Filter Toggle and parchment to Loadout Screen (#3095)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes grain a toggle with it on by default along with fixing a slight
error in a previous pull request and adding parchment to the loadout
menu
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Grain is personal preference and also doesnt look very good on maps like
Desert Town.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
local server testing
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->